Create final_project
When the application is ran, users are faced with a GUI that has an entry box at the top.

Pass the following link to it:
https://edgar.jrc.ec.europa.eu/booklet/EDGARv8.0_FT2022_GHG_booklet_2023_fossilCO2_only.xlsx

The link will be downloaded and have its data passed to a pandas data frame object when 'Get Data' is clicked. Afterwards, the drop down menu should update, allowing users to select which two countries compare in a particular sector in terms of their carbon emissions in metric tons.